use package metadata and refactor function arguments
